+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   methods to the `Node` struct.
 - `is_none`, `is_inherit`, `is_current_color` and `is_aspect_ratio`
   methods to the `Attribute` struct.
+- Implement all `is_*` methods form the `Attribute` struct to `AttributeValue`.
 
 ### Removed
 - `Attribute::visible` field.